+++ mock-api-test-sdk-net80/users/CommonTestConstants.cs
@@ -4,5 +4,8 @@ namespace mock_api_test_sdk_net80
     {
         public const long TEST_USER_ID = 1234567890L;
         public const long TEST_PLAN_ID = 1234567890123456L;
+        public const long TEST_SHEET_ID = 9876543210L;
+        public const long TEST_WORKSPACE_ID = 1122334455L;
+        public const long TEST_REPORT_ID = 2233445566L;
     }
 }