+++ smartsheet-csharp-sdk/main/Smartsheet/Api/Internal/ReportResourcesImpl.cs
@@ -323,6 +323,7 @@ namespace Smartsheet.Api.Internal
         /// <exception cref="AuthorizationException"> if there is any problem with  the REST API authorization (access token) </exception>
         /// <exception cref="ResourceNotFoundException"> if the resource cannot be found </exception>
         /// <exception cref="ServiceUnavailableException"> if the REST API service is not available (possibly due to rate limiting) </exception>
+        /// <exception cref="ArgumentException"> if scopes are empty </exception>
         /// <exception cref="SmartsheetException"> if there is any other error during the operation </exception>
         public void AddReportScope(long reportId, IEnumerable<ReportScopeInclusion> scopes)
         {
@@ -372,6 +373,7 @@ namespace Smartsheet.Api.Internal
         /// <exception cref="AuthorizationException"> if there is any problem with  the REST API authorization (access token) </exception>
         /// <exception cref="ResourceNotFoundException"> if the resource cannot be found </exception>
         /// <exception cref="ServiceUnavailableException"> if the REST API service is not available (possibly due to rate limiting) </exception>
+        /// <exception cref="ArgumentException"> if scopes are empty </exception>
         /// <exception cref="SmartsheetException"> if there is any other error during the operation </exception>
         public void RemoveReportScope(long reportId, IEnumerable<ReportScopeInclusion> scopes)
         {