+++ CHANGELOG.md
@@ -4,6 +4,8 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 ## [X.X.X] - Unreleased
+### Added
+- Add 'Add Report Scope' and 'Remove Report Scope' endpoint support to ReportResources
 
 ## [6.6.4] - 2026-01-19
 ### Fixed